web-ui/requirements.txt
elasticsearch==8.11.0
flask==3.0.0
python-dateutil==2.8.2
Flask-Session==0.5.0
gunicorn==21.2.0
python-dotenv==1.0.1
